refactor(aggregator): extract AggregationPostProcessing component
Move the 7 aggregation pre/post-processing methods into a cohesive
AggregationPostProcessing component (aggregator_postprocessing.py) that
holds the single injected collaborator, aggregation_prompt_processor, by
shared reference from PlaybookAggregator.__init__ — the SAME instance
resolved from the AGGREGATION_PROMPT_PROCESSOR ServiceKey, so the
enterprise EnterpriseAggregationPromptProcessor injection and redaction
Protocol seam are preserved unchanged.

Method bodies moved verbatim. self.aggregation_prompt_processor stays a
facade attribute (run/_generate_playbooks_with_source_clusters read it).
Facade orchestration callers delegate to self._postproc.<m>. Thin facade
delegators are kept only for the three methods that OSS tests invoke by
name (_postprocess_aggregation_output,
_aggregation_prompt_extra_instructions_for_context,
_record_postprocessing_artifacts). No public-surface or behavior change.
